modules/audio-libsoundio: Code gardening

+++ modules/audio-libsoundio/src/mixer.cxx
@@ -7,7 +7,7 @@ namespace Moon {
   void Mixer::mix(struct SoundIoChannelArea *areas, const struct SoundIoChannelLayout &layout, const float sampleRate, unsigned int frames)
   {
     // silence the buffer first
-    for (size_t channel = 0; channel < layout.channel_count; ++channel) {
+    for (int channel = 0; channel < layout.channel_count; ++channel) {
       memset(areas[channel].ptr, 0, areas[channel].step * frames);
     }
     for(auto const& handle: handles) {
@@ -15,8 +15,8 @@ namespace Moon {
     }
 
     // clipping, to avoid overdrive
-    for (int frame = 0; frame < frames; ++frame) {
-      for (size_t channel = 0; channel < layout.channel_count; ++channel) {
+    for (size_t frame = 0; frame < frames; ++frame) {
+      for (int channel = 0; channel < layout.channel_count; ++channel) {
         float* buffer = (float*)(areas[channel].ptr + areas[channel].step * frame);
         float sample = *buffer;
         *buffer = sample > 1.0f ? 1.0f : (sample < -1.0f ? -1.0f : sample);
